requirements.txt
pyspark==3.5.1
pyarrow==16.1.0
pendulum==3.0.0
great_expectations==0.18.14
apache-airflow-providers-apache-spark==4.7.0
apache-airflow-providers-postgres==5.10.0
